refactor: Replace BigInt to Int

+++ src/main/kotlin/application/port/outbound/external/CurrencyAdapter.kt
@@ -2,13 +2,12 @@ package application.port.outbound
 
 import shared.value.Currency
 import java.math.BigDecimal
-import java.math.BigInteger
 
 /**
  * Port interface for currency operations.
  */
 interface CurrencyAdapter {
-    suspend fun convertToSystem(amount: BigDecimal, currency: Currency): BigInteger
+    suspend fun convertToSystem(amount: BigDecimal, currency: Currency): Long
 
-    suspend fun convertFromSystem(amount: BigInteger, currency: Currency): BigDecimal
+    suspend fun convertFromSystem(amount: Long, currency: Currency): BigDecimal
 }
