+++ src/main/kotlin/application/port/outbound/external/PlayerAdapter.kt
@@ -1,7 +1,5 @@
 package application.port.outbound
 
-import java.math.BigInteger
-
 /**
  * Port interface for player operations.
  * Implementations connect to external player service.
@@ -11,5 +9,5 @@ interface PlayerAdapter {
      * Get player's current bet limit.
      * Returns null if no limit is set.
      */
-    suspend fun findCurrentBetLimit(playerId: String): Result<BigInteger?>
+    suspend fun findCurrentBetLimit(playerId: String): Result<Long?>
 }
