+++ src/main/kotlin/application/port/outbound/external/WalletAdapter.kt
@@ -2,7 +2,6 @@ package application.port.outbound
 
 import domain.session.model.Balance
 import shared.value.Currency
-import java.math.BigInteger
 
 /**
  * Port interface for wallet operations.
@@ -22,8 +21,8 @@ interface WalletAdapter {
         playerId: String,
         transactionId: String,
         currency: Currency,
-        realAmount: BigInteger,
-        bonusAmount: BigInteger
+        realAmount: Long,
+        bonusAmount: Long
     ): Result<Balance>
 
     /**
@@ -34,8 +33,8 @@ interface WalletAdapter {
         playerId: String,
         transactionId: String,
         currency: Currency,
-        realAmount: BigInteger,
-        bonusAmount: BigInteger
+        realAmount: Long,
+        bonusAmount: Long
     ): Result<Balance>
 
     /**
