+++ src/main/kotlin/domain/common/error/DomainError.kt
@@ -1,7 +1,5 @@
 package domain.common.error
 
-import java.math.BigInteger
-
 /**
  * Base sealed class for all domain errors.
  * Using sealed class allows exhaustive when expressions and type-safe error handling.
@@ -42,8 +40,8 @@ data class ValidationError(
  */
 data class InsufficientBalanceError(
     val playerId: String,
-    val required: BigInteger,
-    val available: BigInteger,
+    val required: Long,
+    val available: Long,
     override val cause: Throwable? = null
 ) : DomainError("Insufficient balance for player $playerId: required $required, available $available") {
     override val code: String = "INSUFFICIENT_BALANCE"
@@ -54,8 +52,8 @@ data class InsufficientBalanceError(
  */
 data class BetLimitExceededError(
     val playerId: String,
-    val betAmount: BigInteger,
-    val limit: BigInteger,
+    val betAmount: Long,
+    val limit: Long,
     override val cause: Throwable? = null
 ) : DomainError("Bet limit exceeded for player $playerId: bet $betAmount, limit $limit") {
     override val code: String = "BET_LIMIT_EXCEEDED"
