+++ src/main/kotlin/infrastructure/aggregator/onegamehub/adapter/OneGameHubCurrencyAdapter.kt
@@ -3,18 +3,17 @@ package infrastructure.aggregator.onegamehub.adapter
 import application.port.outbound.CurrencyAdapter
 import shared.value.Currency
 import java.math.BigDecimal
-import java.math.BigInteger
 import java.math.RoundingMode
 
 class OneGameHubCurrencyAdapter(private val currencyAdapter: CurrencyAdapter) {
-    suspend fun convertSystemToProvider(amount: BigInteger, currency: Currency): BigInteger {
+    suspend fun convertSystemToProvider(amount: Long, currency: Currency): Long {
         val original = currencyAdapter.convertFromSystem(amount, currency)
 
-        return (original * BigDecimal("100")).toBigInteger()
+        return (original * BigDecimal("100")).toLong()
     }
 
-    suspend fun convertProviderToSystem(amount: BigInteger, currency: Currency): BigInteger {
-        val original = amount.toBigDecimal().divide(BigDecimal("100"), 2, RoundingMode.HALF_UP)
+    suspend fun convertProviderToSystem(amount: Long, currency: Currency): Long {
+        val original = BigDecimal(amount).divide(BigDecimal("100"), 2, RoundingMode.HALF_UP)
 
         return currencyAdapter.convertToSystem(original, currency)
     }
